+++ bank-feeds/src/codatbankfeeds/models/operations/create_source_account.py
@@ -8,24 +8,27 @@ from ..shared import sourceaccount as shared_sourceaccount
 from typing import Optional
 
 
-
 @dataclasses.dataclass
 class CreateSourceAccountRequest:
     company_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'companyId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a company."""
     connection_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'connectionId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a connection."""
     source_account: Optional[shared_sourceaccount.SourceAccount] = dataclasses.field(default=None, metadata={'request': { 'media_type': 'application/json' }})
     
-
 
 
 
 @dataclasses.dataclass
 class CreateSourceAccountResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
     r"""The request made is not valid."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""
     source_account: Optional[shared_sourceaccount.SourceAccount] = dataclasses.field(default=None)
     r"""Success"""